backend/requirements.txt
# Zyntel Backend - Python Dependencies
# Install with: pip install -r requirements.txt --break-system-packages
#
# IMPORTANT - Python version:
# Use Python 3.11 (NOT 3.14) to run these scripts.
# psycopg2-binary does not yet have a prebuilt wheel for Python 3.14,
# so installation will fail or require a C compiler.
# Run with: py -3.11 scripts/data-fetching/fetch_lims_data.py
# Or set Python 3.11 as your default in your .env / npm scripts.

# HTTP requests & web scraping (fetch_lims_data.py)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: PostgreSQL driver (Python scripts that may query DB)
# timeout.py does not use DB or boto3/R2 (single hospital, local logging only).
psycopg2-binary>=2.9.9

# SQL Server driver for direct LIMS DB access (fetch_lims_data.py)
# Use: db (direct DB), scrape (web scraping), or auto (try DB, fallback to scrape)
pyodbc>=5.0.0

# Environment variable loading (both scripts)
python-dotenv>=1.0.0